tests(qa): implementa Fase 1 e 2 (backend) do plano de melhoria de QA (#41)
Fase 1 (hardening): fixtures de mock (worker/oracle) em conftest.py,
--strict-markers, testes de worker.py (core, monitor com time-machine,
wakeup/log flusher, security da allowlist) e oracle_extract.py.

Fase 2 (cobertura, backend): threshold elevado para 82%, testes de
run_task/main_loop, notifications (escalada de cooldown, eviction,
dispatch por canal), circuit breaker Oracle, portfolio_catalog e
scheduler_runtime, auditoria de SessionLocal no client fixture.

Corrige 2 bugs de producao encontrados durante a escrita dos testes:
list_scheduled_jobs quebrava ao ordenar jobs com next_run_time misto
(string formatada + None); make_oracle_retry tipava reset_timeout
como int quando deveria ser float.

Vitest no Dashboard (F2-T2) e diff-cover ficam para uma proxima onda.

+++ lib/python/oracle_retry.py
@@ -13,7 +13,7 @@ F = TypeVar("F", bound=Callable[..., Any])
 
 def make_oracle_retry(
     fail_max: int = 3,
-    reset_timeout: int = 60,
+    reset_timeout: float = 60,
     attempts: int = 3,
     wait_initial: float = 0.1,
     wait_max: float = 5.0,
